cleans up and renames files more logically
