--- quant_trading/frontend/scripts/test.py
@@ -1,124 +1,0 @@
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-import backtrader as bt
-from ..models import Stocks
-import pandas as pd
-import numpy as np
-import time
-from .customDataFeed import dbFeed
-
-
-# Create a Stratey
-#TODO put this in seperate file so we can exchange them easily
-class TestStrategy(bt.Strategy):
-
-    def log(self, txt, dt=None):
-        ''' Logging function for this strategy'''
-        dt = dt or self.datas[0].datetime.date(0)
-        print('%s, %s' % (dt.isoformat(), txt))
-
-    def __init__(self):
-        # Keep a reference to the "close" line in the data[0] dataseries
-        self.dataclose = self.datas[0].close
-
-    def next(self):
-        # Simply log the closing price of the series from the reference
-        self.log('Close, %.2f' % self.dataclose[0])
-
-
-def exec():
-    cerebro = bt.Cerebro()
-    # Add a strategy
-    cerebro.addstrategy(TestStrategy)    
-    
-    # Load data from database
-    df = convert_to_dataframe(Stocks.objects.all(),fields=['open', 'high', 'low', 'close', 'volume', 'date'])
-    df['date'] = df['date'].apply(lambda x: np.datetime64(time.strftime( '%Y-%m-%d', time.gmtime(x))))
-    df.set_index('date', inplace=True)
-
-    
-    data = bt.feeds.PandasData(dataname=df)
-    # Create datafeeda
-    #data = df.to_csv()
-    #data = bt.feeds.AbstractDataBase(Stocks.objects.all())
-    
-    # Add the Data Feed to Cerebro
-    cerebro.adddata(data)
-
-    cerebro.broker.setcash(100000.0)
-
-    print('Starting Portfolio Value: %.2f' % cerebro.broker.getvalue())
-
-    cerebro.run()
-
-    print('Final Portfolio Value: %.2f' % cerebro.broker.getvalue())
-
-
-
-def convert_to_dataframe(qs, fields=None, index=None):
-    """
-    ::param qs is an QuerySet from Django
-    ::fields is a list of field names from the Model of the QuerySet
-    ::index is the preferred index column we want our dataframe to be set to
-
-    Using the methods from above, we can easily build a dataframe
-    from this data.
-    """
-    lookup_fields = get_lookup_fields(qs.model, fields=fields)
-    index_col = None
-    if index in lookup_fields:
-        index_col = index
-    elif "id" in lookup_fields:
-        index_col = 'id'
-    values = qs_to_dataset(qs, fields=fields)
-    df = pd.DataFrame.from_records(values, columns=lookup_fields, index=index_col)
-    return df
-
-def get_lookup_fields(model, fields=None):
-    '''
-    ::param model is a Django model class
-    ::param fields is a list of field name strings.
-    This method compares the lookups we want vs the lookups
-    that are available. It ignores the unavailable fields we passed.
-    '''
-    model_field_names = get_model_field_names(model)
-    if fields is not None:
-        '''
-        we'll iterate through all the passed field_names
-        and verify they are valid by only including the valid ones
-        '''
-        lookup_fields = []
-        for x in fields:
-            if "__" in x:
-                # the __ is for ForeignKey lookups
-                lookup_fields.append(x)
-            elif x in model_field_names:
-                lookup_fields.append(x)
-    else:
-        '''
-        No field names were passed, use the default model fields
-        '''
-        lookup_fields = model_field_names
-    return lookup_fields
-
-def qs_to_dataset(qs, fields=None):
-    '''
-    ::param qs is any Django queryset
-    ::param fields is a list of field name strings, ignoring non-model field names
-    This method is the final step, simply calling the fields we formed on the queryset
-    and turning it into a list of dictionaries with key/value pairs.
-    '''
-
-    lookup_fields = get_lookup_fields(qs.model, fields=fields)
-    return list(qs.values(*lookup_fields))
-
-def get_model_field_names(model, ignore_fields=['content_object']):
-    '''
-    ::param model is a Django model class
-    ::param ignore_fields is a list of field names to ignore by default
-    This method gets all model field names (as strings) and returns a list 
-    of them ignoring the ones we know don't work (like the 'content_object' field)
-    '''
-    model_fields = model._meta.get_fields()
-    model_field_names = list(set([f.name for f in model_fields if f.name not in ignore_fields]))
-    return model_field_names
